code refactoring and improvements
