+++ src/components/Note.vue
@@ -88,7 +88,8 @@ import { emit } from '@nextcloud/event-bus'
 import SyncAlertIcon from 'vue-material-design-icons/SyncAlert'
 
 import { config } from '../config'
-import { fetchNote, refreshNote, saveNote, saveNoteManually, autotitleNote, routeIsNewNote, conflictSolutionLocal, conflictSolutionRemote } from '../NotesService'
+import { fetchNote, refreshNote, saveNote, saveNoteManually, autotitleNote, conflictSolutionLocal, conflictSolutionRemote } from '../NotesService'
+import { routeIsNewNote } from '../Util'
 import TheEditor from './EditorEasyMDE'
 import ThePreview from './EditorMarkdownIt'
 import ConflictSolution from './ConflictSolution'
